pyproject.toml
[tool.poetry]
name = "sofomo-geolocation-api"
version = "0.1.0"
description = ""
authors = ["Michał Freliszka <mfreliszka@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.7.1"
ruff = "^0.9.6"
pytest = "^8.3.4"
httpx = "^0.28.1"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
sqlalchemy-stubs = "^0.4"
loguru = "^0.7.3"
greenlet = "^3.1.1"
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.21.0"
sqlalchemy-utils = "^0.41.2"
syrupy = "^4.8.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
